Usernames with dot at the ending
This is a fix for the redirecting of users that has a dot at the ending. Currently if the user has a dot at the ending (of the username), the user will get redirected to 404 not found page. This commit fixed this by replacing the dot with nothing and there fore redirecting properly the user after login.